[IOT-1911] Make resource/csdk/stack W4 compliant.
These changes include the following:
- Fix W4 warnings under resource/csdk/stack.
- Fix W4 warnings under resource/csdk/stack/test.
- Enable /W4 /WX on resource/csdk/stack.
- Enable /W4 /WX on resource/csdk/stack/test.
- Move disabling warnings from resource/csdk/security
  to build_common/Windows.

Reviewed-on: https://gerrit.iotivity.org/gerrit/17831
Tested-by: jenkins-iotivity <jenkins@iotivity.org>

+++ resource/csdk/stack/src/ocstack.c
@@ -1739,7 +1739,7 @@ void OCHandleResponse(const CAEndpoint_t* endPoint, const CAResponseInfo_t* resp
                     return;
                 }
 
-                for (uint8_t i = start; i < responseInfo->info.numOptions; i++)
+                for (int i = start; i < responseInfo->info.numOptions; i++)
                 {
                     memcpy (&(response->rcvdVendorSpecificHeaderOptions[i-start]),
                             &(responseInfo->info.options[i]), sizeof(OCHeaderOption));
